Per-PU configuration of Hibernate Search
