+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/elasticsearch/test/configuration/IndexedEntity.java
@@ -1,14 +1,27 @@
 package io.quarkus.hibernate.search.elasticsearch.test.configuration;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
 
 /**
  * An indexed entity.
- * <p>
- * Technically this particular class is not a valid entity as it is not marked with @Entity.
- * However, that's not relevant for our tests, and it's easier this way.
  */
+@Entity
 @Indexed
 public class IndexedEntity {
 
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
 }
