+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/elasticsearch/test/configuration/NoConfigNoIndexedEntityTest.java
@@ -13,7 +13,8 @@ public class NoConfigNoIndexedEntityTest {
 
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest().setArchiveProducer(
-            () -> ShrinkWrap.create(JavaArchive.class));
+            () -> ShrinkWrap.create(JavaArchive.class)
+                    .addAsResource("application-nohsearchconfig.properties", "application.properties"));
 
     @Test
     public void testNoConfig() throws SQLException {
