+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/resources/application-multiple-persistence-units.properties
@@ -10,16 +10,28 @@ quarkus.datasource.data2.jdbc.url=jdbc:h2:mem:data2;DB_CLOSE_DELAY=-1
 quarkus.hibernate-orm.dialect=org.hibernate.dialect.H2Dialect
 quarkus.hibernate-orm.database.generation=drop-and-create
 
-quarkus.hibernate-orm."pu1".dialect=org.hibernate.dialect.H2Dialect
-quarkus.hibernate-orm."pu1".database.generation=drop-and-create
-quarkus.hibernate-orm."pu1".datasource=data1
-
-quarkus.hibernate-orm."pu2".dialect=org.hibernate.dialect.H2Dialect
-quarkus.hibernate-orm."pu2".database.generation=drop-and-create
-quarkus.hibernate-orm."pu2".datasource=data2
-
 quarkus.hibernate-search-orm.elasticsearch.version=7
 quarkus.hibernate-search-orm.elasticsearch.hosts=${elasticsearch.hosts:localhost:9200}
 quarkus.hibernate-search-orm.elasticsearch.protocol=${elasticsearch.protocol:http}
 quarkus.hibernate-search-orm.schema-management.strategy=drop-and-create-and-drop
 quarkus.hibernate-search-orm.automatic-indexing.synchronization.strategy=sync
+
+quarkus.hibernate-orm."pu1".dialect=org.hibernate.dialect.H2Dialect
+quarkus.hibernate-orm."pu1".database.generation=drop-and-create
+quarkus.hibernate-orm."pu1".datasource=data1
+
+quarkus.hibernate-search-orm."pu1".elasticsearch.version=7
+quarkus.hibernate-search-orm."pu1".elasticsearch.hosts=${elasticsearch.hosts:localhost:9200}
+quarkus.hibernate-search-orm."pu1".elasticsearch.protocol=${elasticsearch.protocol:http}
+quarkus.hibernate-search-orm."pu1".schema-management.strategy=drop-and-create-and-drop
+quarkus.hibernate-search-orm."pu1".automatic-indexing.synchronization.strategy=sync
+
+quarkus.hibernate-orm."pu2".dialect=org.hibernate.dialect.H2Dialect
+quarkus.hibernate-orm."pu2".database.generation=drop-and-create
+quarkus.hibernate-orm."pu2".datasource=data2
+
+quarkus.hibernate-search-orm."pu2".elasticsearch.version=7
+quarkus.hibernate-search-orm."pu2".elasticsearch.hosts=${elasticsearch.hosts:localhost:9200}
+quarkus.hibernate-search-orm."pu2".elasticsearch.protocol=${elasticsearch.protocol:http}
+quarkus.hibernate-search-orm."pu2".schema-management.strategy=drop-and-create-and-drop
+quarkus.hibernate-search-orm."pu2".automatic-indexing.synchronization.strategy=sync
